Move each game into its own directory
This should keep things organized as we support more games.
