+++ src/games/infinite-recharge/schema.sql
